+++ doc/Changelog.md
@@ -3,7 +3,9 @@
 
 # git master {#master}
 
-* [216]()https://github.com/Eyescale/Lunchbox/pull/216:
+* [217](https://github.com/Eyescale/Lunchbox/pull/217):
+  Implement lunchbox::getRootPath()
+* [216](https://github.com/Eyescale/Lunchbox/pull/216):
   getLibraryPaths: Fixes library path results for OSX and linux
 * [213](https://github.com/Eyescale/Lunchbox/pull/213):
   getExecutablePath: More sensible behaviour for OS X app bundles